Make Location manager implement proper interface
- refactoring of whole class
- removing location will trigger additional concurrency stamp update in all
users that were assigned to this location

+++ ScanApp.Infrastructure/Identity/LocationManagerService.cs
@@ -1,16 +1,18 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using ScanApp.Application.Common.Entities;
 using ScanApp.Application.Common.Helpers.Result;
+using ScanApp.Application.Common.Interfaces;
 using ScanApp.Domain.Entities;
 using ScanApp.Infrastructure.Persistence;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Linq.Expressions;
 using System.Threading.Tasks;
 
 namespace ScanApp.Infrastructure.Identity
 {
-    public class LocationManagerService
+    public class LocationManagerService : ILocationManager
     {
         private readonly IDbContextFactory<ApplicationDbContext> _ctxFactory;
 
@@ -26,27 +28,36 @@ namespace ScanApp.Infrastructure.Identity
             return new Result<List<Location>>(locations);
         }
 
-        public async Task<Result<Location>> GetLocationsByIndex(int index)
+        public async Task<Result<Location>> GetLocationByName(string name)
         {
             await using var ctx = _ctxFactory.CreateDbContext();
-            var location = await ctx.Locations
-                .AsNoTracking()
-                .Where(l => l.Id == index)
-                .SingleOrDefaultAsync()
-                .ConfigureAwait(false);
+            var location = await GetLocationBy(ctx, l => l.Name.Equals(name)).ConfigureAwait(false);
+
+            return location is null
+                ? new Result<Location>(ErrorType.NotFound, $"Location {name} does not exist in database")
+                : new Result<Location>(location);
+        }
+
+        public async Task<Result<Location>> GetLocationById(string index)
+        {
+            await using var ctx = _ctxFactory.CreateDbContext();
+            var location = await GetLocationBy(ctx, l => l.Id.Equals(index)).ConfigureAwait(false);
 
             return location is null
                 ? new Result<Location>(ErrorType.NotFound, $"No location with index of {index}")
                 : new Result<Location>(location);
         }
 
+        private static Task<Location> GetLocationBy(IApplicationDbContext ctx, Expression<Func<Location, bool>> predicate) =>
+            ctx.Locations.AsNoTracking().SingleOrDefaultAsync(predicate);
+
         public Task<Result<Location>> AddNewLocation(Location location)
         {
             _ = location ?? throw new ArgumentNullException(nameof(location));
 
-            return location.Id != 0
-                ? Task.FromResult(new Result<Location>(ErrorType.NotValid, "Custom index value is not supported when adding location - set location index to 0"))
-                : AddNewLocation(location.Name);
+            return string.IsNullOrEmpty(location.Id)
+                ? AddNewLocation(location.Name)
+                : Task.FromResult(new Result<Location>(ErrorType.NotValid, "Custom index value is not supported when adding location - set location index to null or empty string"));
         }
 
         public async Task<Result<Location>> AddNewLocation(string locationName)
@@ -75,22 +86,24 @@ namespace ScanApp.Infrastructure.Identity
         {
             await using var ctx = _ctxFactory.CreateDbContext();
 
-            ctx.Locations.Remove(location);
+            var users = await ctx.UserLocations
+                .Where(o => o.LocationId.Equals(location.Id))
+                .Join(ctx.Users, locations => locations.UserId, user => user.Id,
+                    (_, user) => new ApplicationUser { Id = user.Id, ConcurrencyStamp = user.ConcurrencyStamp })
+                .ToListAsync()
+                .ConfigureAwait(false);
 
-            var users = await ctx.Users
-                .AsNoTracking()
-                .Where(u => u.LocationId == location.Id)
-                .Select(u => u.Id)
-                .ToListAsync();
+            if (await ctx.Locations.SingleOrDefaultAsync(l => l.Id.Equals(location.Id)).ConfigureAwait(false) is not null)
+                ctx.Locations.Remove(location);
 
-            var userEntities = users.ConvertAll(id => new ApplicationUser() { Id = id, LocationId = 0 });
-            ctx.Users.AttachRange(userEntities);
-            userEntities.ForEach(user => ctx.Entry(user).Property(u => u.LocationId).IsModified = true);
+            ctx.AttachRange(users);
+            foreach (var applicationUser in users)
+            {
+                applicationUser.ConcurrencyStamp = Guid.NewGuid().ToString();
+            }
 
-            var change = await ctx.SaveChangesAsync();
-            return change != 0
-                ? new Result<Location>(ResultType.Deleted)
-                : new Result<Location>(ErrorType.Unknown, $"Could not delete {location.Name}");
+            await ctx.SaveChangesAsync().ConfigureAwait(false);
+            return new Result<Location>(ResultType.Deleted);
         }
     }
 }
